feat(gguf): consolidate PRs #135-139, split gguf.rs into modular gguf/ directory
Incorporates 5 PRs from mvkorobkov (MLA metadata, multi-shard reader,
MoE fallback, MLA capability gate, streaming GGUF extract) with fixes:
- 3-digit shard width detection, Q4K MLA guard, detect_gguf_entry dedup
- Split 3,221-line monolith into 7 focused modules (93.5% test coverage)
