+++ crates/larql-vindex/src/extract/streaming/mod.rs
@@ -23,7 +23,7 @@ mod context;
 mod stages;
 mod tensor_io;
 
-use std::path::{Path, PathBuf};
+use std::path::Path;
 
 use crate::config::dtype::StorageDtype;
 use crate::config::types::QuantFormat;
@@ -64,7 +64,7 @@ pub fn build_vindex_streaming(
     // from `model_dir`; for GGUF input the architecture lives in the
     // file metadata, so we open the entry GGUF, build a config.json
     // equivalent, and detect from that.
-    let arch = if let Some(gguf_path) = detect_gguf_entry_for_arch(model_dir)? {
+    let arch = if let Some(gguf_path) = context::detect_gguf_entry(model_dir)? {
         let gguf = larql_models::loading::gguf::GgufFile::open(&gguf_path)
             .map_err(|e| VindexError::Parse(format!("open GGUF for arch detection: {e}")))?;
         let cfg_json = gguf.to_config_json();
@@ -96,9 +96,6 @@ pub fn build_vindex_streaming(
         callbacks,
     )?;
 
-    // The arch-detection helper here mirrors `context::detect_gguf_entry`
-    // (private to that module) — kept inline rather than re-exported
-    // because the use site is a single call. (See note at the helper.)
     ctx.write_gate_vectors()?;
     ctx.write_router_weights()?;
     ctx.write_embeddings()?;
@@ -109,44 +106,4 @@ pub fn build_vindex_streaming(
     ctx.finalize()?;
 
     Ok(())
-}
-
-/// Mirror of `context::detect_gguf_entry` used only for architecture
-/// detection — kept here to avoid making the context-private helper
-/// pub(super) just for one call. See `context.rs::detect_gguf_entry`
-/// for the contract.
-fn detect_gguf_entry_for_arch(model_dir: &Path) -> Result<Option<PathBuf>, VindexError> {
-    if model_dir.is_file()
-        && model_dir.extension().is_some_and(|e| e == "gguf")
-    {
-        return Ok(Some(model_dir.to_path_buf()));
-    }
-    if !model_dir.is_dir() {
-        return Ok(None);
-    }
-    let mut gguf_files: Vec<PathBuf> = std::fs::read_dir(model_dir)?
-        .filter_map(|e| e.ok())
-        .map(|e| e.path())
-        .filter(|p| p.extension().is_some_and(|e| e == "gguf"))
-        .collect();
-    if gguf_files.is_empty() {
-        return Ok(None);
-    }
-    gguf_files.sort();
-    if let Some(shard1) = gguf_files.iter().find(|p| {
-        p.file_name()
-            .and_then(|n| n.to_str())
-            .map(|n| n.contains("-00001-of-"))
-            .unwrap_or(false)
-    }) {
-        return Ok(Some(shard1.clone()));
-    }
-    let mut largest: Option<(u64, PathBuf)> = None;
-    for p in gguf_files {
-        let size = std::fs::metadata(&p).map(|m| m.len()).unwrap_or(0);
-        if largest.as_ref().map_or(true, |(s, _)| size > *s) {
-            largest = Some((size, p));
-        }
-    }
-    Ok(largest.map(|(_, p)| p))
 }
